refactor: move legacy controller
