+++ src/lib/features/feature-toggle/legacy/feature-toggle-legacy-controller.ts
@@ -1,35 +1,35 @@
 /* eslint-disable @typescript-eslint/explicit-module-boundary-types */
 import { Request, Response } from 'express';
-import Controller from '../controller';
-import { extractUsername } from '../../util/extract-user';
-import { NONE, UPDATE_FEATURE } from '../../types/permissions';
-import { IUnleashConfig } from '../../types/option';
-import { IUnleashServices } from '../../types';
-import FeatureToggleService from '../../features/feature-toggle/feature-toggle-service';
-import { featureSchema, querySchema } from '../../schema/feature-schema';
-import { IFeatureToggleQuery } from '../../types/model';
-import FeatureTagService from '../../services/feature-tag-service';
-import { IAuthRequest } from '../unleash-types';
-import { DEFAULT_ENV } from '../../util/constants';
+import Controller from '../../../routes/controller';
+import { extractUsername } from '../../../util/extract-user';
+import { NONE, UPDATE_FEATURE } from '../../../types/permissions';
+import { IUnleashConfig } from '../../../types/option';
+import { IUnleashServices } from '../../../types';
+import FeatureToggleService from '../feature-toggle-service';
+import { featureSchema, querySchema } from '../../../schema/feature-schema';
+import { IFeatureToggleQuery } from '../../../types/model';
+import FeatureTagService from '../../../services/feature-tag-service';
+import { IAuthRequest } from '../../../routes/unleash-types';
+import { DEFAULT_ENV } from '../../../util/constants';
 import {
     featuresSchema,
     FeaturesSchema,
-} from '../../openapi/spec/features-schema';
-import { TagSchema } from '../../openapi/spec/tag-schema';
-import { TagsSchema } from '../../openapi/spec/tags-schema';
-import { serializeDates } from '../../types/serialize-dates';
-import { OpenApiService } from '../../services/openapi-service';
-import { createRequestSchema } from '../../openapi/util/create-request-schema';
+} from '../../../openapi/spec/features-schema';
+import { TagSchema } from '../../../openapi/spec/tag-schema';
+import { TagsSchema } from '../../../openapi/spec/tags-schema';
+import { serializeDates } from '../../../types/serialize-dates';
+import { OpenApiService } from '../../../services/openapi-service';
+import { createRequestSchema } from '../../../openapi/util/create-request-schema';
 import {
     createResponseSchema,
     resourceCreatedResponseSchema,
-} from '../../openapi/util/create-response-schema';
+} from '../../../openapi/util/create-response-schema';
 import {
     emptyResponse,
     getStandardResponses,
-} from '../../openapi/util/standard-responses';
-import { UpdateTagsSchema } from '../../openapi/spec/update-tags-schema';
-import { ValidateFeatureSchema } from '../../openapi/spec/validate-feature-schema';
+} from '../../../openapi/util/standard-responses';
+import { UpdateTagsSchema } from '../../../openapi/spec/update-tags-schema';
+import { ValidateFeatureSchema } from '../../../openapi/spec/validate-feature-schema';
 
 const version = 1;
